ml_service/requirements.txt
# ==========================================================
# AuthDocs ML Service - Requirements
# Cleaned & stable version (CPU only)
# ==========================================================

# ===============================
# Core Deep Learning (CPU only)
# ===============================
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu
torchvision==0.18.1+cpu

# ===============================
# Web Framework
# ===============================
fastapi==0.111.0
uvicorn==0.30.0

# ===============================
# Image Processing & Analysis
# ===============================
opencv-python-headless==4.10.0.84
pillow==10.3.0
numpy==1.26.4
scikit-learn==1.5.1
scikit-image==0.24.0
imageio[pyav]==2.37.2

# ===============================
# OCR & Text Extraction
# ===============================
easyocr==1.7.1
pytesseract==0.3.13

# ===============================
# NLP / Embeddings
# ===============================
transformers==4.44.2
sentence-transformers==3.0.1

# ===============================
# Metadata & File Info
# ===============================
piexif==1.1.3
python-magic==0.4.27

# ===============================
# Miscellaneous Utilities
# ===============================
requests==2.32.3
joblib==1.4.2
